# Decompression speed optimizations