Silesia level 1

| Compiler | Before | After | Gain |
|---|---|---|---|
| gcc-9.2 | 1192.7 MB/s | 1256.7 MB/s | +5.3% |
| clang-9.0.0 | 1083 MB/s | 1203.2 MB/s | +11.1% |

enwik8 level 1


| Compiler | Before | After | Gain |
|---|---|---|---|
| gcc-9.2 | 1035.6 MB/s | 1103.8 MB/s | +6.5% |
| clang-9.0.0 | 923.2 MB/s | 1044.8 MB/s | +13.1% |

enwik8 level 22 (`ZSTD_decompressSequencesLong()`)

| Compiler | Before | After | Gain |
|---|---|---|---|
| gcc-9.2 | 739.2 MB/s | 781.2 MB/s | +5.6% |
| clang-9.0.0 | 694.5 MB/s | 746.3 MB/s | +7.5% |

Streaming decompression of enwik8 and silesia at level 1: `zstd -t enwik8.zst silesia.tar.zst`

| Compiler | Before | After | Gain |
|---|---|---|---|
| gcc-9.2 | 927.3 MB/s | 970.3 MB/s | +4.6%|
| clang-9.0.0 | 847.1 MB/s | 935.5 MB/s | +10.4% |

Measured with `perf stat -r 10 ./build/programs/zstd -t ../enwik8.zst ../silesia.tar.zst` gcc-9 does 6% less instructions and 10% less branches, and clang-9.2 does 15% less instructions and 9% less branches. Each with slightly lower IPC.

I ran the `tests/fuzzer` decompression fuzzers for about an hour, and ran `fuzzer.c` and `zstreamtest.c` for 10K iterations. There aren't any obvious bugs found, but more intensive fuzzer tests may expose some.