@terrelln 2019-11-19
This is ready for review now