@caoyzh 2020-01-22
The optimizations of modifications in “lib/decompress/huf_decompress.c” is not very well in aarch64，it would be good if you could give us some more info about your test environment. Thanks.