@terrelln 2020-01-22
Hi @caoyzh, what platform are you testing on (e.g. raspberry pi or a particular phone)? I did my testing on an Intel i9-9900K. I also tested on a 2018 Macbook Pro and an Intel Skylake processor.

I'm currently looking into mobile performance, so I can look into this once I know what platform you are seeing the regression on.