@terrelln 2020-01-22
@caoyzh what is the raw speed of gcc before/after? Can you compare it to clang?