@Yikun 2020-01-23
@terrelln 
> What platform are you testing on (e.g. raspberry pi or a particular phone)?

Thanks for reply, the platform we are testing is the server with [Kunpeng processer](http://www.hisilicon.com/en/Products/ProductList/Kunpeng) (which is a 64-bit Arm-compatible server CPU released by Huawei Hisilicon).