@caoyzh 2020-01-23
I test it on KungPeng based on armv8.
the raw speed is as follow:
![搜狗截图20200123114019](https://user-images.githubusercontent.com/56909772/72954949-2ff2ef80-3dd5-11ea-9570-ecb86feca50d.png)
the average decompression optimizations of Level 1~19 in clang-9.0.0 is -1.19%