@terrelln 2020-01-23
> Could you give some explanation about the optimization idea？The code is more complex than before, but I’m not very sure if the generated assembly will be simpler. Thanks.

@caoyzh can you please reply in the main thread. I don't seem to be able to view your comment on GitHub, maybe because of a GitHub bug. So I'm replying here.

The optimization is to replace `BIT_reloadDStream()` with `BIT_reloadDStreamFast()`, which has less code, so the loop is smaller. The edge cases are pulled out and handled in the `HUF_decodeStream*()` calls below. The cost is in the `HUF_decompress4X1` loop we now have to check the `endSignal`. The check isn't expensive, and should have been done anyways. The check is already present in the `HUF_decompress4X2` loop.

I've found the compilers to be very picky about this code, and it is tricky to get an incantation that works well for all compilers and architectures. The more architectures we add to the matrix, the harder it gets.

I'm sorry to say that I'm not going to be able to debug the performance on your server. Without access to the hardware it will be impossible. However, I will accept a patch that fixes the performance, as long as it doesn't hurt any of the architecture/compiler combinations that we care about most. Since it is only a 1-1.5% regression, I don't think it is worth having a conditional compilation that has special code for your compiler/architecture combination. If there were a 10% regression I would consider it.

I'm going to put up a patch later today or this week that slightly improves aarch64 performance with clang on mobile. It isn't an optimization really, just a slight rewrite to help guide the compiler. Maybe it will help your performance.

The architectures and compilers we care about most, in order, are:
* x86_64 clang and gcc
* aarch64 and armv7 clang decompression only (phones)
* i386 clang and gcc
* everything else

Of course, we care about correctness on every architecture deeply, and will always fix any correctness bugs.